requirements.txt
Django==3.1.3
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.1.0
python-dotenv==0.19.2
Pillow==9.0.0